refactor tests for ConnectionParams
