+++ test.ts
@@ -1,3 +1,4 @@
 import "./tests/queries.ts";
+import "./tests/connection_params.ts";
 import "./tests/client.ts";
 import "./tests/utils.ts";
